Fix resolution of references
Resolve failed with the following exception when using references:
Exception in thread "main" com.typesafe.config.ConfigException$BugOrBroken: replace in parent not possible

The docs for resolveWith says:
> Note that this method does NOT look in this instance for substitution values.

Which means we can't refer to variabled in the same config.
It also says:
> If you want to do that, you could either merge this instance into
> your value source using Config.withFallback,  or you could resolve
> multiple times with multiple sources (using setAllowUnresolved).

I tried using withFallback, but the same error still occurred.
This might be due to a bug in typesafe config:
https://github.com/lightbend/config/issues/332

This works around that by resolving multiple times as suggested
in the javadoc.